Refactored app/pages/common/Modals/ForkServerModal
